Local

+++ article.go
@@ -1,12 +1,12 @@
 package main
 
 type Article struct {
-	SubId     int64  `json:"subId"`
+	SubId     int    `json:"subId"`
 	Title     string `json:"title"`
 	Content   string `json:"content"`
 	Link      string `json:"link"`
 	Published int64  `json:"published"`
-	Prev      int64  `json:"prev,omitempty"`
+	Prev      int    `json:"prev,omitempty"`
 }
 
 func (p *Article) Size() int {
